--- app/routes/artist.$username._index.tsx
@@ -1,356 +1,0 @@
-import {defer, type LoaderFunctionArgs} from '@shopify/remix-oxygen';
-
-import {
-  Button,
-  Card,
-  Divider,
-  Flex,
-  Grid,
-  Group,
-  rem,
-  Skeleton,
-  Stack,
-  Text,
-  Title,
-} from '@mantine/core';
-import '@mantine/tiptap/styles.css';
-import {Await, Link, Outlet, useLoaderData} from '@remix-run/react';
-import {Suspense, useEffect, useMemo, useState} from 'react';
-
-import {parseGid} from '@shopify/hydrogen';
-import {da} from 'date-fns/locale';
-
-import {format} from 'date-fns';
-import type {
-  ScheduleFragment,
-  TreatmentProductFragment,
-} from 'storefrontapi.generated';
-import {LocationIcon} from '~/components/LocationIcon';
-import {GET_USER_PRODUCTS} from '~/graphql/queries/GetUserProducts';
-import {useUser} from '~/hooks/use-user';
-import {type CustomerScheduleSlot} from '~/lib/api/model';
-import {convertLocations} from '~/lib/convertLocations';
-import {useDuration} from '~/lib/duration';
-import {useTranslations} from '~/providers/Translation';
-
-export async function loader({request, params, context}: LoaderFunctionArgs) {
-  const {username} = params;
-
-  if (!username) {
-    throw new Response('username handler not defined', {
-      status: 404,
-    });
-  }
-
-  const collection = context.storefront.query(GET_USER_PRODUCTS, {
-    variables: {
-      handle: username,
-      filters: [
-        {
-          productMetafield: {
-            namespace: 'system',
-            key: 'type',
-            value: 'product',
-          },
-        },
-        {
-          productMetafield: {
-            namespace: 'booking',
-            key: 'hide_from_profile',
-            value: 'false',
-          },
-        },
-      ],
-    },
-    cache: context.storefront.CacheShort(),
-  });
-
-  return defer({
-    collection,
-  });
-}
-
-export default function ArtistIndex() {
-  const data = useLoaderData<typeof loader>();
-  const {t} = useTranslations();
-  const user = useUser();
-
-  return (
-    <>
-      <Title order={2} fw="600" mb="md">
-        {t('artist_treatments_title')}
-      </Title>
-      <Suspense fallback={<Skeleton height={8} radius="xl" />}>
-        <Await resolve={data.collection}>
-          {({collection}) => {
-            const collections =
-              collection?.products.nodes.reduce((collections, product) => {
-                if (!collections[product.productType]) {
-                  collections[product.productType] = [];
-                }
-                collections[product.productType].push(product);
-                return collections;
-              }, {} as Record<string, TreatmentProductFragment[]>) || {};
-            return (
-              <Grid gutter="xl">
-                <Grid.Col span={{base: 12, md: 8}}>
-                  <Stack gap="xl">
-                    {Object.keys(collections).map((key) => (
-                      <Stack key={key} gap="sm">
-                        <Title order={4} fw="600">
-                          {key}
-                        </Title>
-                        <Stack gap="md">
-                          {collections[key].map((product) => (
-                            <ArtistProduct key={product.id} product={product} />
-                          ))}
-                        </Stack>
-                      </Stack>
-                    ))}
-                  </Stack>
-                </Grid.Col>
-                <Grid.Col span={{base: 12, md: 4}}>
-                  <Stack gap="md">
-                    {user.schedules
-                      ?.sort((a, b) => {
-                        const aContainsMonday =
-                          a.slots?.value?.includes('monday');
-                        const bContainsMonday =
-                          b.slots?.value?.includes('monday');
-
-                        if (aContainsMonday && !bContainsMonday) {
-                          return -1;
-                        } else if (!aContainsMonday && bContainsMonday) {
-                          return 1;
-                        } else {
-                          return 0;
-                        }
-                      })
-                      .map((schedule) => (
-                        <Schedule key={schedule.handle} schedule={schedule} />
-                      ))}
-                  </Stack>
-                </Grid.Col>
-              </Grid>
-            );
-          }}
-        </Await>
-        <Outlet />
-      </Suspense>
-    </>
-  );
-}
-
-async function wordToColor(word: string) {
-  const encoder = new TextEncoder();
-  const data = encoder.encode(word);
-
-  const hashBuffer = await crypto.subtle.digest('SHA-256', data);
-
-  const hashArray = Array.from(new Uint8Array(hashBuffer));
-  const hashHex = hashArray
-    .map((b) => b.toString(16).padStart(2, '0'))
-    .join('');
-
-  const color = `#${hashHex.substring(0, 6)}`;
-
-  return color;
-}
-
-function Schedule({schedule}: {schedule: ScheduleFragment}) {
-  const {t} = useTranslations();
-  const [scheduleColor, setScheduleColor] = useState<string>('#fff');
-
-  useEffect(() => {
-    const fetchColor = async () => {
-      const color = await wordToColor(schedule.handle || '');
-      setScheduleColor(color);
-    };
-
-    fetchColor();
-  }, [schedule]);
-
-  const slots = JSON.parse(
-    schedule.slots?.value || '',
-  ) as CustomerScheduleSlot[];
-
-  return (
-    <Card withBorder radius="md">
-      <Card.Section bg={scheduleColor} px="md" py="sm">
-        <Title order={3} c="white">
-          {schedule.name?.value}
-        </Title>
-      </Card.Section>
-      <Divider mb="md" />
-      <Title order={4} mb="xs">
-        {t('artist_openingtime')}
-      </Title>
-      <Stack gap={rem(3)}>
-        {slots?.map((slot) => {
-          return (
-            <Group key={slot.day}>
-              <Text>{t(slot.day)}</Text>
-              {slot.intervals.map((interval) => {
-                return (
-                  <Group key={interval.from + interval.to}>
-                    <Text>
-                      {format(timeToDate(interval.from, new Date()), 'HH:mm', {
-                        locale: da,
-                      })}{' '}
-                      -{' '}
-                      {format(timeToDate(interval.to, new Date()), 'HH:mm', {
-                        locale: da,
-                      })}
-                    </Text>
-                  </Group>
-                );
-              })}
-            </Group>
-          );
-        })}
-      </Stack>
-      <Card.Section>
-        <Divider my="md" />
-      </Card.Section>
-      <Stack gap="xs">
-        {schedule.locations?.references?.nodes.map((location) => (
-          <Group key={location.handle}>
-            <LocationIcon
-              location={{locationType: location.locationType?.value as any}}
-              color={scheduleColor || '#000000'}
-            />
-            {location.locationType?.value === 'destination' ||
-            location.locationType?.value === 'virtual' ? (
-              <>
-                {
-                  location.fullAddress?.value?.split(',')[
-                    location.fullAddress?.value?.split(',').length - 1
-                  ]
-                }
-              </>
-            ) : (
-              <>{location.fullAddress?.value}</>
-            )}
-          </Group>
-        ))}
-      </Stack>
-    </Card>
-  );
-}
-
-export function ArtistProduct({product}: {product: TreatmentProductFragment}) {
-  const durationToTime = useDuration();
-  const [scheduleColor, setScheduleColor] = useState<string>('#fff');
-  const {t} = useTranslations();
-
-  useEffect(() => {
-    const fetchColor = async () => {
-      const color = await wordToColor(
-        product.scheduleId?.reference?.handle || '',
-      );
-      setScheduleColor(color);
-    };
-
-    fetchColor();
-  }, [product.scheduleId]);
-
-  const productId = parseGid(product?.id).id;
-  const locations = convertLocations(product.locations?.references?.nodes);
-  const variant = product.variants.nodes[0];
-
-  const discountString = useMemo(() => {
-    if (
-      variant.compareAtPrice?.amount &&
-      variant.compareAtPrice?.amount !== '0.0'
-    ) {
-      const discountAmount =
-        parseInt(variant.compareAtPrice?.amount) -
-        parseInt(variant.price.amount);
-      const discountPercentage = Math.abs(
-        (discountAmount / parseInt(variant.compareAtPrice?.amount)) * 100,
-      );
-      return `${t('discount')} ${discountPercentage.toFixed(0)}%`;
-    }
-    return null;
-  }, [t, variant.compareAtPrice?.amount, variant.price.amount]);
-
-  return (
-    <Card
-      key={product.handle}
-      withBorder
-      component={Link}
-      radius="md"
-      data-testid={`service-item-${productId}`}
-      to={`treatment/${product.handle}`}
-    >
-      <Grid>
-        <Grid.Col span={8}>
-          <Flex direction="column" gap="xs">
-            <div>
-              <Group gap="xs">
-                <Title
-                  order={2}
-                  size="md"
-                  fw={500}
-                  lts=".5px"
-                  data-testid={`service-title-${productId}`}
-                >
-                  {product.title}
-                </Title>
-                {locations
-                  .map((l) => ({
-                    locationType: l.locationType,
-                  }))
-                  .filter(
-                    (value, index, self) =>
-                      index ===
-                      self.findIndex(
-                        (t) => t.locationType === value.locationType,
-                      ),
-                  )
-                  .map((location) => (
-                    <LocationIcon
-                      key={location.locationType}
-                      location={location}
-                      color={scheduleColor}
-                    />
-                  ))}
-              </Group>
-
-              <Text
-                c="dimmed"
-                size="sm"
-                data-testid={`service-duration-text-${productId}`}
-              >
-                {durationToTime(product.duration?.value || 0)}
-              </Text>
-            </div>
-
-            <Group>
-              <Text size="sm">
-                {product.options?.value ? t('from') : ''} {variant.price.amount}{' '}
-                kr
-              </Text>
-              {discountString ? <Text size="sm">{discountString}</Text> : null}
-            </Group>
-          </Flex>
-        </Grid.Col>
-        <Grid.Col span={4}>
-          <Flex justify="flex-end" align="center" h="100%">
-            <Button variant="outline" c="black" color="gray.4" radius="lg">
-              {t('artist_booktime')}
-            </Button>
-          </Flex>
-        </Grid.Col>
-      </Grid>
-    </Card>
-  );
-}
-
-// Helper function to convert time string to Date object
-function timeToDate(time: string, date: Date): Date {
-  const [hour, minute] = time.split(':').map(Number);
-  const newDate = new Date(date);
-  newDate.setUTCHours(hour, minute, 0, 0);
-  return newDate;
-}
